rever factories move, add copyRelations to relation/functions file
